enable go-get support by renaming main source to radi
